Changed the method of counting attempts
Now the attempts are counted using the i in the for loop, rather than using a seperate variable.
This change has both positive and negative aspects as it makes the programme logic less understandable, but clears up the code and uses less variables to achieve the same result.
In such a small project however the positives of the new approach outshine the negatives.